requirements.txt
langchain>=0.1.20
langchain-community>=0.0.24
langgraph>=0.0.34
openai>=1.30.1
duckduckgo-search>=4.0
tiktoken>=0.7.0
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
yfinance
psycopg2
sqlalchemy
redis